Observations List

API endpoint that represents a list of observations.

ee_observations(
  page = NULL,
  page_size = 1000,
  country = "United States",
  state_province = NULL,
  county = NULL,
  kingdom = NULL,
  phylum = NULL,
  order = NULL,
  clss = NULL,
  family = NULL,
  genus = NULL,
  scientific_name = NULL,
  kingdom__exact = NULL,
  phylum__exact = NULL,
  order__exact = NULL,
  clss__exact = NULL,
  family__exact = NULL,
  genus__exact = NULL,
  scientific_name__exact = NULL,
  remote_id = NULL,
  collection_code = NULL,
  source = NULL,
  min_date = NULL,
  max_date = NULL,
  georeferenced = FALSE,
  bbox = NULL,
  exclude = NULL,
  extra = NULL,
  quiet = FALSE,
  progress = FALSE,
  foptions = list()
)

Arguments

page
Page number

page_size
The number of observations per page returned by the query. The package default is 250 observations. The API default, however, is 10. Setting this number higher will require fewer calls (i.e. pages) but a large size may also fail or timeout.

country
country name

state_province
description needed.

county
California county. See data(california_counties)

kingdom
kingdom name

phylum
phylum name

order
order name

clss
class name

family
family name

genus
genus name.

scientific_name
A full scientific name

kingdom__exact
exact kingdom name

phylum__exact
exact phylum name

order__exact
exact order name

clss__exact
class name

family__exact
exact family name

genus__exact
exact genus name

scientific_name__exact
exact scientific name

remote_id
remote ID

collection_code
collections code

source
data source. See ee_sources

min_date
Lower date bound. Most be in format Y-m-D h:m:s. This package uses lubridate's ymd_hms function to format dates.

max_date
upper date bound. Most be in format Y-m-D h:m:s. This package uses lubridate's ymd_hms function to format dates.

georeferenced
Default is FALSE. Set to TRUE to return only georeferenced records.

bbox
Set a bounding box for your search. Use format bbox=-124,32,-114,42. Order is min Longitude , min Latitude , max Longitude , max Latitude. Use http://boundingbox.klokantech.com/ this website to quickly grab a bounding box (set format to csv on lower right)

exclude
Default is NULL. Pass a list of fields to exclude.

extra
Default is NULL. Pass a list of extra fields to be returned. Additional fields include: # "id", "record", "source", "remote_resource", "begin_date", "end_date", # "collection_code", "institution_code", "state_province", "county", # "last_modified", "original_id", "geometry", "coordinate_uncertainty_in_meters" # "md5", "scientific_name", "observation_type", "date_precision", "locality" # "earliest_period_or_lowest_system", "latest_period_or_highest_system", "kingdom" # "phylum", "clss", "order", "family", "genus", "specific_epithet", # "infraspecific_epithet", "minimum_depth_in_meters", "maximum_depth_in_meters", # "maximum_elevation_in_meters", "minimum_elevation_in_meters", "catalog_number" # "preparations", "sex", "life_stage", "water_body", "country", "individual_count", "associated_resources"

quiet
Default is FALSE. Set to TRUE to suppress messages.

progress
Progress bar is set to TRUE by default. Use FALSE especially when calling functions inside documents.

foptions
A list of additional arguments for httr. There is no reason to use this argument except for debugging purposes.

Value

data.frame

Examples

# vulpes <- ee_observations(genus = "vulpes")
if (FALSE) {
# pinus <- ee_observations(scientific_name = "Pinus", page_size = 100)
# lynx_data <- ee_observations(genus = "Lynx")
# Georeferenced data only
# lynx_data <- ee_observations(genus = "Lynx", georeferenced = TRUE)
# animalia <- ee_observations(kingdom = "Animalia")
# Artemisia <- ee_observations(scientific_name = "Artemisia douglasiana")
# asteraceae <- ee_observationss(family = "asteraceae")
# vulpes <- ee_observations(genus = "vulpes")
# Anas <- ee_observations(scientific_name = "Anas cyanoptera", page = "all")
# loons <- ee_observations(scientific_name = "Gavia immer", page = "all")
# plantae <- ee_observations(kingdom = "plantae")
# chordata <- ee_observations(phylum = "chordata")
# Class is clss since the former is a reserved keyword in SQL.
# aves <- ee_observations(clss = "aves")
# You can also exclude fields using the exclude field
#  aves <- ee_observations(clss = "aves", exclude = "remote_resource")
#  aves <- ee_observations(clss = "aves", exclude = "source,remote_resource")
#  or request additional fields using extra
#  aves <- ee_observations(clss = "aves", extra = "kingdom,genus")
# aves <- ee_observations(clss = "aves", bbox = '-124,32,-114,42')
# aves <- ee_observations(clss = "aves", county = "Alameda county")
}